pyproject.toml
# Copyright 2025 IO Club
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "g03-multipurpose-qqbot"
version = "1.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "qq-botpy>=1.2.1",
]

authors = [{ name = "celet-ff-io" }]
license = "Apache-2.0"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
include = ["g03mpqb*"]

[tool.poetry]
packages = [{ include = "g03mpqb", from = "src" }]
